+++ src/spyglass/spikesorting/merge.py
@@ -1,5 +1,3 @@
-import uuid
-
 import datajoint as dj
 
 from spyglass.spikesorting.imported import ImportedSpikeSorting  # noqa: F401
@@ -8,6 +6,7 @@ from spyglass.spikesorting.spikesorting_curation import (  # noqa: F401
 )
 from spyglass.spikesorting.v1.curation import CurationV1  # noqa: F401
 from spyglass.utils.dj_merge_tables import _Merge
+from spyglass.utils.dj_mixin import SpyglassMixin
 
 schema = dj.schema("spikesorting_merge")
 
@@ -21,21 +20,21 @@ class SpikeSortingOutput(_Merge):
     source: varchar(32)
     """
 
-    class CurationV1(dj.Part):  # noqa: F811
+    class CurationV1(SpyglassMixin, dj.Part):  # noqa: F811
         definition = """
         -> master
         ---
         -> CurationV1
         """
 
-    class ImportedSpikeSorting(dj.Part):  # noqa: F811
+    class ImportedSpikeSorting(SpyglassMixin, dj.Part):  # noqa: F811
         definition = """
         -> master
         ---
         -> ImportedSpikeSorting
         """
 
-    class CuratedSpikeSorting(dj.Part):  # noqa: F811
+    class CuratedSpikeSorting(SpyglassMixin, dj.Part):  # noqa: F811
         definition = """
         -> master
         ---